Api

Public Functions

JSServe.App — Type

App(handler)

calls handler with the session and the http request object. f is expected to return a valid DOM object, so something renderable by jsrender, e.g. DOM.div.

JSServe.CodeEditor — Method

CodeEditor(language::String; initial_source="", theme="chrome", editor_options...)

Defaults for editor_options:

(
    autoScrollEditorIntoView = true,
    copyWithEmptySelection = true,
    wrapBehavioursEnabled = true,
    useSoftTabs = true,
    enableMultiselect = true,
    showLineNumbers = false,
    fontSize = 16,
    wrap = 80,
    mergeUndoDeltas = "always"
)

JSServe.NoServer — Type

We don't serve files and include anything directly as raw bytes. Interpolating the same asset many times, will only upload the file to JS one time though.

JSServe.Session — Type

A web session with a user

JSServe.WebSocketConnection — Method

handles a new websocket connection to a session

JSServe.Page — Method

Page(;
    offline=false, exportable=true,
    connection::Union{Nothing, FrontendConnection}=nothing,
    server_config...
)

A Page can be used for resetting the JSServe state in a multi page display outputs, like it's the case for Pluto/IJulia/Documenter. For Documenter, the page needs to be set to exportable=true, offline=true, but doesn't need to, since Page defaults to the most common parameters for known Packages. Exportable has the effect of inlining all data & js dependencies, so that everything can be loaded in a single HTML object. offline=true will make the Page not even try to connect to a running Julia process, which makes sense for the kind of static export we do in Documenter. For convenience, one can also pass additional server configurations, which will directly get put into configure_server!(;server_config...). Have a look at the docs for configure_server! to see the parameters.

JSServe.configure_server! — Method

configure_server!(;
        listen_url::String=SERVER_CONFIGURATION.listen_url[],
        listen_port::Integer=SERVER_CONFIGURATION.listen_port[],
        forwarded_port::Integer=listen_port,
        external_url=nothing,
        content_delivery_url=nothing
    )

Configures the parameters for the automatically started server.

Parameters:

* listen_url=SERVER_CONFIGURATION.listen_url[]
    The address the server listens to.
    must be 0.0.0.0, 127.0.0.1, ::, ::1, or localhost.
    If not set differently by an ENV variable, will default to 127.0.0.1

* listen_port::Integer=SERVER_CONFIGURATION.listen_port[],
    The Port to which the default server listens to
    If not set differently by an ENV variable, will default to 9284

* forwarded_port::Integer=listen_port,
    if port gets forwarded to some other port, set it here!

* external_url=nothing
    The url from which the server is reachable.
    If served on "127.0.0.1", this will default to http://localhost:forwarded_port
    if listen_url is "0.0.0.0", this will default to http://$(Sockets.getipaddr()):forwarded_port
    so that the server is reachable inside the local network.
    If the server should be reachable from some external dns server,
    this needs to be set here.

* content_delivery_url=nothing
    You can server files from another server.
    Make sure any file referenced from Julia is reachable at
    content_delivery_url * "/the_file"

JSServe.evaljs — Method

evaljs(session::Session, jss::JSCode)

Evaluate a javascript script in session.

JSServe.evaljs_value — Method

evaljs_value(session::Session, js::JSCode)

Evals js code and returns the jsonified value. Blocks until value is returned. May block indefinitely, when called with a session that doesn't have a connection to the browser.

JSServe.linkjs — Method

linkjs(session::Session, a::Observable, b::Observable)

for an open session, link a and b on the javascript side. This will also Link the observables in Julia, but only as long as the session is active.

JSServe.onjs — Method

onjs(session::Session, obs::Observable, func::JSCode)

Register a javascript function with session, that get's called when obs gets a new value. If the observable gets updated from the JS side, the calling of func will be triggered entirely in javascript, without any communication with the Julia session.

Private Functions

JSServe.Asset — Type

Represent an asset stored at an URL. We try to always have online & local files for assets

JSServe.FrontendConnection — Type

Inteface for FrontendConnection

struct MyConnection <: FrontendConnection
end

Needs to have a constructor with 0 arguments:

MyConnection()

Needs to overload Base.write for sending binary data

Base.write(connection::MyConnection, bytes::AbstractVector{UInt8})

Needs to implement isopen to indicate status of connection

Base.isopen(c::MyConnection)

Setup connection will be called before rendering any dom with session. The return value will be inserted into the DOM of the rendered App and can be used to do the JS part of opening the connection.

JSServe.setup_connection(session::Session{IJuliaConnection})::Union{JSCode, Nothing}

One can overload use_parent_session, to turn on rendering dom objects inside sub-sessions while keeping one parent session managing the connection alive. This is handy for IJulia/Pluto, since the parent session just needs to be initialized one time and can stay active and globally store objects used multiple times across doms.

JSServe.use_parent_session(::Session{MyConnection}) = false/false

JSServe.JSCode — Type

Javascript code that supports interpolation of Julia Objects. Construction of JSCode via string macro:

jsc = js"console.log($(some_julia_variable))"

This will decompose into:

jsc.source == [JSString("console.log("), some_julia_variable, JSString(""")]

JSServe.JSException — Method

Creates a Julia exception from data passed to us by the frondend!

JSServe.JSString — Type

The string part of JSCode.

JSServe.JSUpdateObservable — Type

Functor to update JS part when an observable changes. We make this a Functor, so we can clearly identify it and don't sent any updates, if the JS side requires to update an Observable (so we don't get an endless update cycle)

JSServe.Table — Type

A simple wrapper for types that conform to the Tables.jl Table interface, which gets rendered nicely!

JSServe.add_cached! — Method

add_cached!(create_cached_object::Function, session::Session, message_cache::Dict{String, Any}, key::String)

Checks if key is already cached by the session or it's root session (we skip any child session between root -> this session). If not cached already, we call create_cached_object to create a serialized form of the object corresponding to key and cache it. We return nothing if already cached, or the serialized object if not cached. We also handle the part of adding things to the message_cache from the serialization context.

JSServe.contextual_eval — Method

contextual_eval(parent, expr)

"Evals" expression without eval by only allowing getfield + getindex expressions

contextual_eval(context, :(lala.blalba)) == context.lala.blabla
contextual_eval(context, :(lala.blalba[1])) == context.lala.blabla[1]
contextual_eval(context, :(julia_func())) == error

JSServe.dependency_path — Method

dependency_path(paths...)

Path to serve downloaded dependencies

JSServe.export_standalone — Method

export_standaloneexport_standalone(
    app::App, folder::String;
    clear_folder=false, write_index_html=true,
    absolute_urls=false, content_delivery_url="file://" * folder * "/",
    single_html=false)

Exports the app defined by app::Application with all its assets to folder. Will write the main html out into folder/index.html. Overwrites all existing files! If this gets served behind a proxy, set absolute_urls=true and set content_delivery_url to your proxy url. If clear_folder=true all files in folder will get deleted before exporting again! single_html=true will write out a single html instead of writing out JS depencies as separate files.

JSServe.getextension — Method

getextension(path)

Get the file extension of the path. The extension is defined to be the bit after the last dot, excluding any query string.

Examples

julia> JSServe.getextension("foo.bar.js")
"js"
julia> JSServe.getextension("https://my-cdn.net/foo.bar.css?version=1")
"css"

Taken from WebIO.jl

JSServe.is_online — Method

is_online(path)

Determine whether or not the specified path is a local filesystem path (and not a remote resource that is hosted on, for example, a CDN).

JSServe.jsrender — Method

jsrender([::Session], x::Any)

Internal render method to create a valid dom. Registers used observables with a session And makes sure the dom only contains valid elements. Overload jsrender(::YourType) To enable putting YourType into a dom element/div. You can also overload it to take a session as first argument, to register messages with the current web session (e.g. via onjs).

JSServe.on_document_load — Method

on_document_load(session::Session, js::JSCode)

executes javascript after document is loaded

JSServe.onload — Method

onload(session::Session, node::Node, func::JSCode)

calls javascript func with node, once node has been displayed.

JSServe.page_html — Method

page_html(session::Session, html_body)

Embeds the html_body in a standalone html document!

JSServe.process_message — Method

process_message(session::Session, bytes::AbstractVector{UInt8})

Handles the incoming websocket messages from the frontend. Messages are expected to be gzip compressed and packed via MsgPack.

JSServe.register_asset_server! — Method

register_asset_server!(condition::Function, ::Type{<: AbstractAssetServer})

Registers a new asset server type. condition is a function that should return nothing, if the asset server type shouldn't be used, and an initialized asset server object, if the conditions are right. E.g. The JSServe.NoServer be used inside an IJulia notebook so it's registered like this:

register_asset_server!(NoServer) do
    if isdefined(Main, :IJulia)
        return NoServer()
    end
    return nothing
end

The last asset server registered takes priority, so if you register a new connection last in your Package, and always return it, You will overwrite the connection type for any other package. If you want to force usage temporary, try:

force_asset_server(YourAssetServer()) do
    ...
end
# which is the same as:
force_asset_server!(YourAssetServer())
...
force_asset_server!()

JSServe.register_connection! — Method

register_connection!(condition::Function, ::Type{<: FrontendConnection})

Registers a new Connection type.

condition is a function that should return nothing, if the connection type shouldn't be used, and an initialized Connection, if the conditions are right. E.g. The IJulia connection should only be used inside an IJulia notebook so it's registered like this:

register_connection!(IJuliaConnection) do
    if isdefined(Main, :IJulia)
        return IJuliaConnection()
    end
    return nothing
end

The last connection registered take priority, so if you register a new connection last in your Package, and always return it, You will overwrite the connection type for any other package. If you want to force usage temporary, try:

force_connection(YourConnectionType()) do
    ...
end
# which is the same as:
force_connection!(YourConnectionType())
...
force_connection!()

JSServe.replace_expressions — Method

replace_expressions(markdown, context)

Replaces all expressions inside markdown savely, by only supporting getindex/getfield expression that will index into context

JSServe.string_to_markdown — Method

string_to_markdown(source::String, context; eval_julia_code=false)

Replaces all interpolation expressions inside markdown savely, by only supporting getindex/getfield expression that will index into context. You can eval Julia code blocks by setting eval_julia_code to a Module, into which the code gets evaluated!

JSServe.update_nocycle! — Method

Update the value of an observable, without sending changes to the JS frontend. This will be used to update updates from the forntend.

Sockets.send — Method

send(session::Session; attributes...)

Send values to the frontend via JSON for now